introduce useProposalDetailsPage hook & use one css class for all buttons in voting overview
