+++ app/components/views/ProposalDetails/ProposalDetails.module.css
@@ -57,7 +57,7 @@
   height: 60px;
 }
 
-.castVoteButton {
+.overviewVotingButton {
   padding: 6px 18px;
   position: relative;
   top: -4px;
